// src/public-api/model/createSellerDto.ts
/**
 * MISCore Shop API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateSellerDto { 
    shopName: string;
    shopDescription?: string | null;
    businessLicenseNumber?: string | null;
    taxId?: string | null;
    contactEmail?: string | null;
    contactPhone?: string | null;
    address?: string | null;
    city?: string | null;
    state?: string | null;
    postalCode?: string | null;
    country?: string | null;
}